+++ frontend/src/components/Login/KakaoLogout.jsx
@@ -1,0 +1,14 @@
+import { KaKaoLoginImg } from '../../assets';
+import useKakaoLogout from '../../hooks/Login/useKakaoLogout';
+
+const KakaoLogout = () => {
+  const { logoutClickHandler } = useKakaoLogout();
+
+  return (
+    <button onClick={logoutClickHandler}>
+      <img src={KaKaoLoginImg} />
+    </button>
+  );
+};
+
+export default KakaoLogout;
