+++ frontend/src/components/Login/LoginForm.jsx
@@ -1,15 +1,7 @@
-import KaKaoLogin from './KaKaoLogin';
+import KakaoLogin from './KakaoLogin';
 import * as S from './LoginStyledComp.style';
-import { KAKAO_AUTHCODE_URL } from '../../constants/auth';
 
 const LoginForm = () => {
-  const kakaoClickHandler = () => {
-    //카카오로그인 -> 리다이렉트 with auth code
-    console.log('카카오로그인 클릭');
-    console.log(KAKAO_AUTHCODE_URL);
-    window.location.assign(KAKAO_AUTHCODE_URL);
-  };
-
   return (
     <S.LoginContainer>
       <div //임시로 넣어둠
@@ -17,7 +9,7 @@ const LoginForm = () => {
       >
         일반회원 회원가입...로그인
       </div>
-      <KaKaoLogin clickHandler={kakaoClickHandler} />
+      <KakaoLogin />
     </S.LoginContainer>
   );
 };
